Add support for setting FontConfig.Name
IMGUI will synthesize a diagnostic name for fonts added from files
(eg "foo.ttf, 16px") if there isn't a name provided in the FontConfig.

There isn't an equivalent synthetic name for fonts added from memory,
so enable the name to be set on the FontConfig.

The names are visible in the Dear Imgui Demo window
under Configuration > Style, in the Fonts drop-list.